requirements.txt
python-dotenv
Flask
SQLAlchemy
click
requests
numpy
pandas
Jinja2
marshmallow
pytest
pytest-cov
coverage
pymongo
dnspython